==> Step1_Define_SCT_Array_Intensity_FinnGen.R <==

### Define sex chromosome aneuploidy (SCA) by using sex chromosome LRR/BAF generated from MoChA pipeline 
## male: XXY, XYY, and XXYY
## female: XXX and XO (note the definition of XO need to be improved, but it's fine for the current project since the PheWAS only focus on trisomy)



#------------------------------------------------------------------------------------
#---Set environments-----------------------------------------------------------------
#------------------------------------------------------------------------------------

setwd("/home/aoxliu/SCA/SCA_FinnGenR10")
# install.packages("data.table")
library(data.table)
library(dplyr)
library(ggplot2)
'%!in%' <- function(x,y)!('%in%'(x,y))

jobid <- "3c5e5b25-df8a-4fd2-af8c-5a2cb1cad5eb"
pfx <- "FinnGenR10Affymetrix"




########################################################################
#               Filter stats info generated by MoChA                   #
########################################################################

## Read in stats data -----------------
df_stats <- data.frame(read.table(paste0("../dsge-cromwell/mocha/",jobid,"/call-mocha_stats_tsv/",pfx,".stats.tsv"), sep="\t", header=T)) %>% 
      mutate(sample_id=as.character(sample_id), computed_gender=as.character(computed_gender)) 
dim(df_stats)   # 347,611     23


## QC for genotyping quality (call_rate lower than 0.97; baf_auto greater than 0.03) -----------------
df_stats <- df_stats %>% filter(call_rate>=0.97 & baf_auto<=0.03)
dim(df_stats)   # 345,923     23


# Adjust sex chromosome LRR median by autosome LRR median -----------------
df_stats <- df_stats %>% mutate(x_nonpar_adj_lrr_median=x_nonpar_lrr_median-lrr_median, 
                                y_nonpar_adj_lrr_median=y_nonpar_lrr_median-lrr_median,
                                mt_adj_lrr_median=mt_lrr_median-lrr_median)


## N of individuals by computed gender -----------------
df_stats <- df_stats %>% mutate(computed_gender=ifelse(computed_gender=="F", "Female", computed_gender)) %>% 
                         mutate(computed_gender=ifelse(computed_gender=="M", "Male", computed_gender)) %>% 
                         mutate(computed_gender=ifelse(computed_gender=="U", "Undetermined", computed_gender)) 

df_stats %>% group_by(computed_gender) %>% count()




########################################################################
#               Filter calls info generated by MoChA                   #
########################################################################

## Read in calls and only keep those from sex chromosome -----------------
df_calls <- data.frame(read.table(paste0("../dsge-cromwell/mocha/",jobid,"/call-mocha_calls_tsv/",pfx,".calls.tsv"), sep="\t", header=T)) %>% 
      mutate(sample_id=as.character(sample_id), computed_gender=as.character(computed_gender)) 
dim(df_calls)   # 484,025     22


## only keep calls from sex chromosome -----------------
df_calls <- df_calls %>% filter(chrom=="chrX")
dim(df_calls)   # 127,173     22


## N of individuals by computed gender -----------------
df_calls <- df_calls %>% mutate(computed_gender=ifelse(computed_gender=="F", "Female", computed_gender)) %>% 
                         mutate(computed_gender=ifelse(computed_gender=="M", "Male", computed_gender)) %>% 
                         mutate(computed_gender=ifelse(computed_gender=="U", "Undetermined", computed_gender)) 

df_calls %>% group_by(computed_gender) %>% count()


## filter calls for each sex -----------------
df_calls <- df_calls %>% filter((computed_gender=="Male" & length > 2e6) | (computed_gender=="Female" & length > 1e8) | n_hets==0)
dim(df_calls)   # 93,940    22

df_calls %>% group_by(computed_gender) %>% count()


## rename columns and only keep useful columns -----------------
df_calls <- df_calls %>% rename(sex_cnv_length="length", sex_cnv_bdev="bdev", sex_n_hets=n_hets) %>% 
                         mutate(sex_cnv=1) %>% 
                         select(sample_id, sex_cnv_length, sex_cnv_bdev, sex_n_hets, sex_cnv)
dim(df_calls)   # 93,940     5


## add sex chromosome cnv info to df_stats -----------------
df_stats <- df_stats %>% left_join(df_calls, by="sample_id") %>% 
      mutate(sex_cnv=ifelse(is.na(sex_cnv), 0, sex_cnv),
             sex_cnv_bdev=ifelse(is.na(sex_cnv_bdev), 0, sex_cnv_bdev),
             sex_cnv_length=ifelse(is.na(sex_cnv_length), 0, sex_cnv_length) )
dim(df_stats)  # 346,259     30




########################################################################
#                  Define XXY (male with an extra X)                   #
########################################################################

df_stats <- df_stats %>% mutate(SCA=ifelse(y_nonpar_adj_lrr_median>-0.5 & y_nonpar_adj_lrr_median<0.5 & x_nonpar_adj_lrr_median>-0.125 & x_nonpar_adj_lrr_median<0.125, "XXY", computed_gender))
df_stats <- df_stats %>% mutate(SCA=ifelse(SCA=="Female", "XX", SCA), SCA=ifelse(SCA=="Male", "XY", SCA))
df_stats %>% distinct(sample_id, SCA) %>% group_by(SCA) %>% count() %>% arrange(desc(n))




########################################################################
#                Define XYY (male with an extra Y)                     #
########################################################################

df_stats <- df_stats %>% mutate(SCA=ifelse((y_nonpar_adj_lrr_median>0.25 & sex_cnv_bdev>0.125)|(y_nonpar_adj_lrr_median>0.375 & sex_cnv_bdev==0), "XYY", SCA)) 
df_stats %>% distinct(sample_id, SCA) %>% group_by(SCA) %>% count() %>% arrange(desc(n))




########################################################################
#          Define XXYY (male with an extra X and an extra Y)           #
########################################################################

df_stats <- df_stats %>% mutate(SCA=ifelse((y_nonpar_adj_lrr_median>0.25 & x_nonpar_adj_lrr_median>-0.125), "XXYY", SCA)) 
df_stats %>% distinct(sample_id, SCA) %>% group_by(SCA) %>% count() %>% arrange(desc(n))




########################################################################
#                Define XXX (female with an extra Y)                   #
########################################################################

df_stats <- df_stats %>% mutate(SCA=ifelse(x_nonpar_adj_lrr_median>0.125 & sex_cnv_bdev>0.125, "XXX", SCA)) 
df_stats %>% distinct(sample_id, SCA) %>% group_by(SCA) %>% count() %>% arrange(desc(n))




################################################################################
#      Define XO (female with one X missing), many of them are mosaic          #
################################################################################

# df_stats <- df_stats %>% mutate(SCA=ifelse(x_nonpar_adj_lrr_median< -0.125 & sex_n_hets==0 & sex_cnv==1 & computed_gender!="Male", "XO", SCA)) # this way is incorrect
df_stats <- df_stats %>% mutate(SCA=ifelse(x_nonpar_adj_lrr_median< -0.125 & SCA=="XX", "XO", SCA)) 
df_stats %>% distinct(sample_id, SCA) %>% group_by(SCA) %>% count() %>% arrange(desc(n))




########################################################################
#                          Write out                                   #
########################################################################

dat_sca <- df_stats %>% rename(FINNGENID="sample_id") %>% distinct(FINNGENID, SCA)
write.table(dat_sca, "finngen_R10_sca.tsv", append=F, quote=F, sep="\t", row.names=F, col.names=T)




########################################################################
#                  Plot (N for the plot is from R11)                   #
########################################################################

## chrX_LRR and chrY_LRR for all individuals -----------------
p <- ggplot(df_stats, aes(x=x_nonpar_adj_lrr_median, y=y_nonpar_adj_lrr_median, color=SCA)) +
  geom_point(data=df_stats, aes(shape=SCA, size=SCA, alpha=SCA)) +
  geom_point(aes(shape=SCA, size=SCA, alpha=SCA)) +
  scale_x_continuous('X nonPAR median LRR (autosome corrected)') + 
  scale_y_continuous('Y nonPAR median LRR (autosome corrected)') +
  scale_color_manual('Sex chromosome', values=c("XY"="#00798c", "XX"="#d1495b", "Undetermined"="#8d96a3", "XXY"="#f37735", "XYY"="#7CAE00", "XXYY"="#2e4057", "XXX"="#AA4371", "XO"="pink"), 
                                       labels=c("XY"="XY (N=167,938)", "XX"="XX (N=222,479)", "Undetermined"="Undetermined (N=2,445)", "XXY"="XXY (N=263)", "XYY"="XYY (N=216)", "XXYY"="XXYY (N=1)", "XXX"="XXX (N=168)", "XO"="XO (N=289)")) + 
  scale_shape_manual('Sex chromosome', values=c("XY"=20, "XX"=20, "Undetermined"=20, "XXY"=2, "XYY"=2, "XXYY"=2, "XXX"=0, "XO"=0), 
                                       labels=c("XY"="XY (N=167,938)", "XX"="XX (N=222,479)", "Undetermined"="Undetermined (N=2,445)", "XXY"="XXY (N=263)", "XYY"="XYY (N=216)", "XXYY"="XXYY (N=1)", "XXX"="XXX (N=168)", "XO"="XO (N=289)")) +
  scale_alpha_manual('Sex chromosome', values=c("XY"=0.1, "XX"=0.1, "Undetermined"=0.1, "XXY"=1, "XYY"=1, "XXYY"=1, "XXX"=1, "XO"=1), 
                                       labels=c("XY"="XY (N=167,938)", "XX"="XX (N=222,479)", "Undetermined"="Undetermined (N=2,445)", "XXY"="XXY (N=263)", "XYY"="XYY (N=216)", "XXYY"="XXYY (N=1)", "XXX"="XXX (N=168)", "XO"="XO (N=289)")) +
  scale_size_manual('Sex chromosome', values=c("XY"=0.2, "XX"=0.2, "Undetermined"=0.2, "XXY"=2, "XYY"=2, "XXYY"=2, "XXX"=2, "XO"=2), 
                                       labels=c("XY"="XY (N=167,938)", "XX"="XX (N=222,479)", "Undetermined"="Undetermined (N=2,445)", "XXY"="XXY (N=263)", "XYY"="XYY (N=216)", "XXYY"="XXYY (N=1)", "XXX"="XXX (N=168)", "XO"="XO (N=289)")) +
  theme_bw(base_size=16) + 
  theme(legend.position='right')
print(p)
ggsave("FinnGenR10_Affymetrix_SCA_by_LRR_BAF.png", p, width=12, height=9)



## chrX_LRR and chrX_BAF for females -----------------
p <- ggplot(df_stats %>% filter(SCA %in% c("XX","XXX","XO")), aes(x=sex_cnv_bdev, y=x_nonpar_adj_lrr_median, color=SCA)) +
  geom_point(data=df_stats %>% filter(SCA %in% c("XX","XXX","XO")), aes(shape=SCA, size=SCA, alpha=SCA)) +
  geom_point(aes(shape=SCA, size=SCA, alpha=SCA)) +
  scale_x_continuous('X BAF deviation') + 
  scale_y_continuous('X nonPAR median LRR (autosome corrected)') +
  scale_color_manual('Sex chromosome', values=c("XX"="#d1495b", "XXX"="#AA4371", "XO"="pink"), 
                                       labels=c("XX"="XX (N=222,479)", "XXX"="XXX (N=168)", "XO"="XO (N=289)")) + 
  scale_shape_manual('Sex chromosome', values=c("XX"=20, "XXX"=0, "XO"=0), 
                                       labels=c("XX"="XX (N=222,479)", "XXX"="XXX (N=168)", "XO"="XO (N=289)")) +
  scale_alpha_manual('Sex chromosome', values=c("XX"=0.1, "XXX"=1, "XO"=1), 
                                       labels=c("XX"="XX (N=222,479)", "XXX"="XXX (N=168)", "XO"="XO (N=289)")) +
  scale_size_manual('Sex chromosome',  values=c("XX"=0.2, "XXX"=2, "XO"=2), 
                                       labels=c("XX"="XX (N=222,479)", "XXX"="XXX (N=168)", "XO"="XO (N=289)")) +
  theme_bw(base_size=16) + 
  theme(legend.position='right')
print(p)
ggsave("FinnGenR10_Affymetrix_SCA_by_LRR_BAF_females.png", p, width=12, height=9)



## chrX_LRR and chrX_BAF for males -----------------
p <- ggplot(df_stats %>% filter(SCA %in% c("XY","XXY","XYY","XXYY")), aes(x=sex_cnv_bdev, y=y_nonpar_adj_lrr_median, color=SCA)) +
  geom_point(data=df_stats %>% filter(SCA %in% c("XY","XXY","XYY","XXYY")), aes(shape=SCA, size=SCA, alpha=SCA)) +
  geom_point(aes(shape=SCA, size=SCA, alpha=SCA)) +
  scale_x_continuous('X BAF deviation') + 
  scale_y_continuous('X nonPAR median LRR (autosome corrected)') +
  scale_x_continuous('PAR1/PAR2 BAF deviation') + 
  scale_y_continuous('Y nonPAR median LRR (autosome corrected)') +
  scale_color_manual('Sex chromosome', values=c("XY"="#00798c", "XXY"="#f37735", "XYY"="#7CAE00", "XXYY"="#2e4057"), 
                                       labels=c("XY"="XY (N=167,938)", "XXY"="XXY (N=263)", "XYY"="XYY (N=216)", "XXYY"="XXYY (N=1)")) + 
  scale_shape_manual('Sex chromosome', values=c("XY"=20, "XXY"=2, "XYY"=2, "XXYY"=2), 
                                       labels=c("XY"="XY (N=167,938)", "XXY"="XXY (N=263)", "XYY"="XYY (N=216)", "XXYY"="XXYY (N=1)")) + 
  scale_alpha_manual('Sex chromosome', values=c("XY"=0.1, "XXY"=1, "XYY"=1, "XXYY"=1), 
                                       labels=c("XY"="XY (N=167,938)", "XXY"="XXY (N=263)", "XYY"="XYY (N=216)", "XXYY"="XXYY (N=1)")) + 
  scale_size_manual('Sex chromosome', values=c("XY"=0.2, "XXY"=2, "XYY"=2, "XXYY"=2), 
                                       labels=c("XY"="XY (N=167,938)", "XXY"="XXY (N=263)", "XYY"="XYY (N=216)", "XXYY"="XXYY (N=1)")) + 
  theme_bw(base_size=16) + 
  theme(legend.position='right')
print(p)
ggsave("FinnGenR10_Affymetrix_SCA_by_LRR_BAF_males.png", p, width=12, height=9)
